requirements.txt
azure-functions==1.8.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
pdf2image==1.16.0
Pillow==9.0.0
pycparser==2.21
pytesseract==0.3.8
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rpy2==3.4.5
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
